+++ workers/tests/integrationtests/test_api_indicator.py
@@ -22,7 +22,7 @@ class TestApiIndicator(unittest.TestCase):
     def setUp(self):
         self.client = TestClient(app)
 
-        self.indicator_name = "Minimal"
+        self.indicator_name = "minimal"
         self.layer_key = "minimal"
         # Heidelberg
         self.dataset = "regions"
@@ -46,9 +46,9 @@ class TestApiIndicator(unittest.TestCase):
     def test_indicator_dataset_default_fid_field(self):
         parameters = {
             "name": self.indicator_name,
-            "layerKey": self.layer_key,
+            "layer_key": self.layer_key,
             "dataset": self.dataset,
-            "featureId": self.feature_id,
+            "feature_id": self.feature_id,
         }
         for response in (
             self.client.get(ENDPOINT + "?" + urlencode(parameters)),
@@ -60,10 +60,10 @@ class TestApiIndicator(unittest.TestCase):
     def test_indicator_dataset_custom_fid_field(self):
         parameters = {
             "name": self.indicator_name,
-            "layerKey": self.layer_key,
+            "layer_key": self.layer_key,
             "dataset": self.dataset,
-            "featureId": self.feature_id,
-            "fidField": self.fid_field,
+            "feature_id": self.feature_id,
+            "fid_field": self.fid_field,
         }
         for response in (
             self.client.get(ENDPOINT + "?" + urlencode(parameters)),
@@ -75,10 +75,10 @@ class TestApiIndicator(unittest.TestCase):
     def test_indicator_dataset_custom_fid_field_2(self):
         parameters = {
             "name": self.indicator_name,
-            "layerKey": self.layer_key,
+            "layer_key": self.layer_key,
             "dataset": self.dataset,
-            "featureId": "Heidelberg",
-            "fidField": "name",
+            "feature_id": "Heidelberg",
+            "fid_field": "name",
         }
         for response in (
             self.client.get(ENDPOINT + "?" + urlencode(parameters)),
@@ -90,9 +90,9 @@ class TestApiIndicator(unittest.TestCase):
     def test_indicator_dataset_invalid(self):
         parameters = {
             "name": self.indicator_name,
-            "layerKey": self.layer_key,
+            "layer_key": self.layer_key,
             "dataset": "foo",
-            "featureId": self.feature_id,
+            "feature_id": self.feature_id,
         }
         for response in (
             self.client.get(ENDPOINT + "?" + urlencode(parameters)),
@@ -107,12 +107,12 @@ class TestApiIndicator(unittest.TestCase):
         for parameters in (
             {
                 "name": self.indicator_name,
-                "layerKey": "building_count",
+                "layer_key": "building_count",
                 "dataset": "regions",
             },
             {
                 "name": self.indicator_name,
-                "layerKey": "building_count",
+                "layer_key": "building_count",
                 "feature_id": "3",
             },
         ):
@@ -127,8 +127,8 @@ class TestApiIndicator(unittest.TestCase):
     @oqt_vcr.use_cassette()
     def test_indicator_include_svg_true(self):
         url = (
-            "/indicator?name={0}&layerKey={1}&dataset={2}"
-            "&featureId={3}&fidField={4}&includeSvg={5}".format(
+            "/indicator?name={0}&layer_key={1}&dataset={2}"
+            "&feature_id={3}&fid_field={4}&include_svg={5}".format(
                 self.indicator_name,
                 self.layer_key,
                 self.dataset,
@@ -144,8 +144,8 @@ class TestApiIndicator(unittest.TestCase):
     @oqt_vcr.use_cassette()
     def test_indicator_include_svg_false(self):
         url = (
-            "/indicator?name={0}&layerKey={1}&dataset={2}"
-            "&featureId={3}&fidField={4}&includeSvg={5}".format(
+            "/indicator?name={0}&layer_key={1}&dataset={2}"
+            "&feature_id={3}&fid_field={4}&include_svg={5}".format(
                 self.indicator_name,
                 self.layer_key,
                 self.dataset,
@@ -161,8 +161,8 @@ class TestApiIndicator(unittest.TestCase):
     @oqt_vcr.use_cassette()
     def test_indicator_include_svg_default(self):
         url = (
-            "/indicator?name={0}&layerKey={1}&dataset={2}"
-            "&featureId={3}&fidField={4}".format(
+            "/indicator?name={0}&layer_key={1}&dataset={2}"
+            "&feature_id={3}&fid_field={4}".format(
                 self.indicator_name,
                 self.layer_key,
                 self.dataset,
@@ -178,9 +178,9 @@ class TestApiIndicator(unittest.TestCase):
     def test_indicator_invalid_layer(self):
         parameters = {
             "name": self.indicator_name,
-            "layerKey": "amenities",
+            "layer_key": "amenities",
             "dataset": "regions",
-            "featureId": "3",
+            "feature_id": "3",
         }
         for response in (
             self.client.get(ENDPOINT + "?" + urlencode(parameters)),
@@ -193,8 +193,8 @@ class TestApiIndicator(unittest.TestCase):
     @oqt_vcr.use_cassette()
     def test_indicator_include_html_true(self):
         url = (
-            "/indicator?name={0}&layerKey={1}&dataset={2}"
-            "&featureId={3}&fidField={4}&includeHtml={5}".format(
+            "/indicator?name={0}&layer_key={1}&dataset={2}"
+            "&feature_id={3}&fid_field={4}&include_html={5}".format(
                 self.indicator_name,
                 self.layer_key,
                 self.dataset,
@@ -210,8 +210,8 @@ class TestApiIndicator(unittest.TestCase):
     @oqt_vcr.use_cassette()
     def test_indicator_include_html_false(self):
         url = (
-            "/indicator?name={0}&layerKey={1}&dataset={2}"
-            "&featureId={3}&fidField={4}&includeHtml={5}".format(
+            "/indicator?name={0}&layer_key={1}&dataset={2}"
+            "&feature_id={3}&fid_field={4}&include_html={5}".format(
                 self.indicator_name,
                 self.layer_key,
                 self.dataset,
@@ -227,8 +227,8 @@ class TestApiIndicator(unittest.TestCase):
     @oqt_vcr.use_cassette()
     def test_indicator_include_html_default(self):
         url = (
-            "/indicator?name={0}&layerKey={1}&dataset={2}"
-            "&featureId={3}&fidField={4}".format(
+            "/indicator?name={0}&layer_key={1}&dataset={2}"
+            "&feature_id={3}&fid_field={4}".format(
                 self.indicator_name,
                 self.layer_key,
                 self.dataset,
@@ -242,7 +242,7 @@ class TestApiIndicator(unittest.TestCase):
 
     @oqt_vcr.use_cassette()
     def test_indicator_flatten_default(self):
-        url = "/indicator?name={0}&layerKey={1}&dataset={2}&featureId={3}".format(
+        url = "/indicator?name={0}&layer_key={1}&dataset={2}&feature_id={3}".format(
             self.indicator_name,
             self.layer_key,
             self.dataset,
@@ -257,8 +257,8 @@ class TestApiIndicator(unittest.TestCase):
     @oqt_vcr.use_cassette()
     def test_indicator_flatten_true(self):
         url = (
-            "/indicator?name={0}&layerKey={1}&dataset={2}"
-            "&featureId={3}&flatten={4}".format(
+            "/indicator?name={0}&layer_key={1}&dataset={2}"
+            "&feature_id={3}&flatten={4}".format(
                 self.indicator_name,
                 self.layer_key,
                 self.dataset,
@@ -273,8 +273,8 @@ class TestApiIndicator(unittest.TestCase):
     @oqt_vcr.use_cassette()
     def test_indicator_flatten_false(self):
         url = (
-            "/indicator?name={0}&layerKey={1}&dataset={2}"
-            "&featureId={3}&flatten={4}".format(
+            "/indicator?name={0}&layer_key={1}&dataset={2}"
+            "&feature_id={3}&flatten={4}".format(
                 self.indicator_name,
                 self.layer_key,
                 self.dataset,
@@ -289,7 +289,7 @@ class TestApiIndicator(unittest.TestCase):
 
     @oqt_vcr.use_cassette()
     def test_indicator_include_data_default(self):
-        url = "/indicator?name={0}&layerKey={1}&dataset={2}&featureId={3}".format(
+        url = "/indicator?name={0}&layer_key={1}&dataset={2}&feature_id={3}".format(
             self.indicator_name,
             self.layer_key,
             self.dataset,
@@ -302,8 +302,8 @@ class TestApiIndicator(unittest.TestCase):
     @oqt_vcr.use_cassette()
     def test_indicator_include_data_true(self):
         url = (
-            "/indicator?name={0}&layerKey={1}&dataset={2}"
-            "&featureId={3}&includeData={4}".format(
+            "/indicator?name={0}&layer_key={1}&dataset={2}"
+            "&feature_id={3}&include_data={4}".format(
                 self.indicator_name,
                 self.layer_key,
                 self.dataset,
@@ -318,8 +318,8 @@ class TestApiIndicator(unittest.TestCase):
     @oqt_vcr.use_cassette()
     def test_indicator_include_data_false(self):
         url = (
-            "/indicator?name={0}&layerKey={1}&dataset={2}"
-            "&featureId={3}&includeData={4}".format(
+            "/indicator?name={0}&layer_key={1}&dataset={2}"
+            "&feature_id={3}&include_data={4}".format(
                 self.indicator_name,
                 self.layer_key,
                 self.dataset,